requirements.txt
langchain
langchain_openai
langchain_community
openai
python-dotenv
PyPDF2 
faiss-cpu
streamlit
streamlit_pdf_viewer 
watchdog
pypdf
huggingface_hub
rapidocr-onnxruntime
pymupdf
